requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
Jinja2==3.1.2
sqlalchemy==2.0.23
bcrypt==4.0.1
PyJWT==2.8.0
python-multipart==0.0.6
email-validator==2.1.0.post1